Move query functions to Database class and update scraper pbd CLI entrypoint
