+++ proboards_scraper/database/__init__.py
@@ -1,3 +1,2 @@
 from .database import Database
-from .query import query_boards, query_users
 from .schema import Base, Board, Category, Moderator, Post, Thread, User
